+++ atn-atm-sdk/src/protocols/mod.rs
@@ -1,0 +1,6 @@
+#[derive(Default)]
+pub struct Protocols {
+    pub message_pickup: message_pickup::MessagePickup,
+}
+
+pub mod message_pickup;